refactor: refactor project view following the mvc architecture
