--- src/main/kotlin/lotto/core/LottoController.kt
@@ -1,28 +1,0 @@
-package lotto.core
-
-import lotto.ui.InputView
-import lotto.ui.ResultView
-
-class LottoController {
-    fun run() {
-        val purchaseAmount = InputView.purchaseAmountInput()
-        val lottoMachine = LottoLogic()
-
-        val numberOfTickets = lottoMachine.calculateNumberOfTickets(purchaseAmount)
-        val tickets = lottoMachine.generateTickets(numberOfTickets)
-
-        val winningLottoList = InputView.winningNumbersInput()
-        val winningLotto = lottoMachine.generateWinningLotto(winningLottoList)
-
-        val bonusNumber = InputView.bonusNumberInput(winningLottoList)
-
-        val result = WinningLogic(tickets, winningLotto, bonusNumber).determineWinningTickets()
-
-        val statistics = Statistics(result, purchaseAmount)
-
-        ResultView.displayNumberOfTickets(numberOfTickets)
-        ResultView.displayTickets(tickets)
-        ResultView.displayWinningRanks(result)
-        ResultView.displayWinningRate(statistics.calculateRate())
-    }
-}
