+++ src/test/kotlin/lotto/core/LottoLogicTest.kt
@@ -1,5 +1,6 @@
 package lotto.core
 
+import lotto.model.LottoLogic
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.ValueSource
