+++ src/test/kotlin/lotto/core/LottoTest.kt
@@ -1,5 +1,6 @@
 package lotto.core
 
+import lotto.model.Lotto
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.assertThrows
 import org.junit.jupiter.params.ParameterizedTest
